+++ changelog.md
@@ -22,6 +22,7 @@
 * Introduced the has file type filter.
 * Fixed bug where the backend would fail hard when an invalid integer was provided as
   'before_id' when moving a row by introducing a decorator to validate query parameters.
+* Fixed bug where copying a cell containing a null value resulted in an error.
 
 ## Released (2021-08-11)